Split clean_air.rs into separate test files by concern
Move shared test utilities into tests/common/mod.rs and split the 10
tests into lookup_tests.rs, fib_tests.rs, and femtocairo_tests.rs for
better navigability and separation of concerns.
